Fix 'tripple sharps' issue when building scales and add more types of scales.
Current code errors when rendering the Pentatonic scales as the number of
sharps builds up/exceeds permitted amount. I believe that this is because
the code progressively builds the scale from the last note used.

This patches code so that it builds using 'root + offset' for each note
in the scale. Tripple sharp issue no longer occurs.